session: split code
