+++ planner/core/casetest/BUILD.bazel
@@ -12,10 +12,6 @@ go_test(
         "plan_test.go",
         "point_get_plan_test.go",
         "predicate_simplification_test.go",
-        "rule_derive_topn_from_window_test.go",
-        "rule_inject_extra_projection_test.go",
-        "rule_join_reorder_test.go",
-        "rule_result_reorder_test.go",
         "stats_test.go",
         "tiflash_selection_late_materialization_test.go",
         "window_push_down_test.go",
@@ -26,12 +22,8 @@ go_test(
     shard_count = 50,
     deps = [
         "//domain",
-        "//expression",
-        "//expression/aggregation",
         "//parser",
-        "//parser/ast",
         "//parser/model",
-        "//parser/mysql",
         "//parser/terror",
         "//planner/core",
         "//planner/core/internal",
@@ -44,10 +36,8 @@ go_test(
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
-        "//types",
         "//util",
         "//util/hint",
-        "//util/mock",
         "//util/plancodec",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
