+++ planner/core/casetest/rule/BUILD.bazel
@@ -1,0 +1,35 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "rule_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "rule_derive_topn_from_window_test.go",
+        "rule_inject_extra_projection_test.go",
+        "rule_join_reorder_test.go",
+        "rule_result_reorder_test.go",
+    ],
+    data = glob(["testdata/**"]),
+    flaky = True,
+    shard_count = 20,
+    deps = [
+        "//domain",
+        "//expression",
+        "//expression/aggregation",
+        "//parser/ast",
+        "//parser/model",
+        "//parser/mysql",
+        "//planner/core/internal",
+        "//sessionctx/variable",
+        "//testkit",
+        "//testkit/testdata",
+        "//testkit/testmain",
+        "//testkit/testsetup",
+        "//types",
+        "//util/mock",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
