+++ planner/core/casetest/rule/rule_derive_topn_from_window_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package casetest
+package rule
 
 import (
 	"testing"
@@ -22,6 +22,8 @@ import (
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/testkit/testdata"
 )
+
+type Input []string
 
 // Rule should bot be applied for TiKV.
 func TestPushDerivedTopnNegative(t *testing.T) {
