+++ planner/core/casetest/rule/rule_join_reorder_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package casetest
+package rule
 
 import (
 	"testing"
