+++ planner/core/casetest/rule/testdata/join_reorder_suite_out.json
